Fix race caused by blindly destroying widgets
When calling the _update_services method we where blindly destroying all
ServiceItem widgets and then drawing new ones. This was both inelegant and
prone to casuing races.
This huge commit, while perhaps no more elegant, keeps the widgets around along
with a list of services and adds new widgetry for each new service we see when
the manager updates.
The list, upon having an item added, will re-order itself based on the services
position in the list of services returned by connman.