Refresh the service list each time it changes.
Carrick was keeping a cached copy of the services list returned by gconnman
and trying to do comparisons on the objects within and only perform miminal
updates on the service list. Unfortuntately there's no reliabel way for
carrick and gconnman to compare service items as what is essentially the
same service could be returned from connman with an entirely different
d-bus object path. We now take a clean copy of the list and refresh the UI
each time we receive the services-changed signal.