CarrickPane: Don't show hidden networks
ConnMan 0.22(ish) changed the way hidden SSID's are presented. Now the name
property will be a string of spaces of varying length depending on AP.
This is a temporary fix to not display services which have a name consisting
entirely of spaces. Once the daemon is fixed we can probably remove this code.
This fixes MB#4284