CarrickPane: Remove border from scrolled views viewport
High on the list of designer issues with carrick was the amount of border
lines around service items (one for each item and one for the list).
The border on the scrolled view was added by the
gtk_scrolled_view_add_with_viewport convenience function. By creating the
viewport ourselves we can set the shadow type on the viewport appropriately.
Much prettier.